backend/pyproject.toml
[tool.poetry]
name = "avicable-backend"
version = "0.0.1"
description = "overhead cables and birdlife - backend"
authors = ["LPO AuRA <auvergne-rhone-alpes@lpo.fr>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = ">=3.10,<4"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.6"
dj-sinp-nomenclatures = "^1.3.3"
djangorestframework-gis = "^1.0"
django-filter = "^23.2"
djoser = "^2.2.0"
django-polymorphic = "^3.1.0"
django-rest-polymorphic = "^0.1.10"
django-cors-headers = "^4.1.0"
django = "^4.2.7"
pillow = "^10.0.0"
docutils = "^0.20.1"
django-debug-toolbar = "^4.2.0"
setuptools = "^75.5.0"
gunicorn = "^23.0.0"
dj-sinp-organisms = "^1.4.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"
pre-commit = "^3.3.3"
django-extensions = "^3.2.3"
django-fixture-magic = "^0.1.5"
pygraphviz = "^1.11"
pydotplus = "^2.0.2"
coverage = "^7.2.7"
django-debug-toolbar = "^4.2.0"
pylint-django = "^2.5.5"
drf-typescript-generator = {git = "https://github.com/lpoaura/drf-typescript-generator"}
drf-yasg = "^1.21.7"


[tool.isort]
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
include = '\.pyi?$'
line-length = 79
target-version = ['py38']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
